driver: use commanded segment time.
Instead of overriding, use commanded segment time to configure the segment
duration.

Uses (in addition to normal registers):

 -  R[3] as segment time cache
 -  R[4] as segment CNT cache

Note: this also changes the driver to only accept trajectories that have
their starting point equal to the 'current joint position' of the robot.
Trajectories for which this is not the case will not be executed, and a
'-2' (Illegal Start State) error will be displayed on the TP.